api_yamdb/api/views.py
from django.shortcuts import get_object_or_404
from rest_framework import viewsets, filters
from django_filters.rest_framework import DjangoFilterBackend
from reviews.models import Category, Genre, Title
from api.filters import TitleFilter
from api.mixins import CreateListDestroyMixins
from api.permissions import IsAdminUserOrReadOnly, CategorySrictGetRequest
from api.serializers import (GenreSerializer,
                             CategorySerializer,
                             TitleSerializer)


class GenreViewSet(CreateListDestroyMixins):
    queryset = Genre.objects.all()
    serializer_class = GenreSerializer
    permission_classes = (IsAdminUserOrReadOnly,)
    lookup_field = 'slug'
    filter_backends = (filters.SearchFilter,)
    search_fields = ('name',)


class CategoryViewSet(CreateListDestroyMixins):
    queryset = Category.objects.all()
    serializer_class = CategorySerializer
    permission_classes = (CategorySrictGetRequest, )
    lookup_field = 'slug'
    filter_backends = (filters.SearchFilter,)
    search_fields = ('name',)


class TitleViewSet(viewsets.ModelViewSet):
    queryset = Title.objects.all()
    serializer_class = TitleSerializer
    permission_classes = (IsAdminUserOrReadOnly,)
    filter_backends = (DjangoFilterBackend,)
    filterset_class = TitleFilter

    def perform_create(self, serializer):
        category = get_object_or_404(
            Category, slug=self.request.data.get('category')
        )
        genre = Genre.objects.filter(
            slug__in=self.request.data.getlist('genre')
        )
        serializer.save(category=category, genre=genre)

    def perform_update(self, serializer):
        self.perform_create(serializer)
