requirements.txt
requests==2.26.0
Django==3.2
djangorestframework-simplejwt==4.8.0
djangorestframework==3.12.4
django-filter==23.2
djoser==2.1.0
flake8-broken-line==0.6.0
flake8-isort==5.0.3
flake8-return==1.2.0
flake8==5.0.4
gunicorn==20.0.4
pep8-naming==0.13.3
psycopg2-binary==2.8.6
PyJWT==2.1.0
pytest==6.2.4
pytest-django==4.4.0
pytest-pythonpath==0.7.3
pytz==2020.1
requests==2.26.0
sqlparse==0.3.1
